restructure finance folder
